REFAC(26): Refatorando o modulo de views.
close #26
