+++ jornada_milhas/core/views/__init__.py
@@ -1,0 +1,9 @@
+from .destination import lc_destination, rud_destination  # noqa: F401
+from .post import lc_post, post_home, rud_post  # noqa: F401
+
+all = (
+    "post_home",
+    "lc_post",
+    "rud_post" "lc_destination",
+    "rud_destination",
+)
