+++ jornada_milhas/core/views/post.py
@@ -3,8 +3,8 @@ from rest_framework.parsers import MultiPartParser
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from jornada_milhas.core.models import Destination, Post
-from jornada_milhas.core.serializer import DestinationSerializer, PostSerializer
+from jornada_milhas.core.models import Post
+from jornada_milhas.core.serializer import PostSerializer
 
 
 class RetrieveUpdateDestroyPost(RetrieveUpdateDestroyAPIView):
@@ -65,58 +65,6 @@ class PostHome(APIView):
         return Response(data={"results": serialiazer.data})
 
 
-class RetrieveUpdateDestroyDestination(RetrieveUpdateDestroyAPIView):
-    queryset = Destination.objects.all()
-    serializer_class = DestinationSerializer
-    parser_classes = [MultiPartParser]
-
-    def get(self, request, *args, **kwargs):
-        """
-        Lendo um destino pelo `id`.
-        """
-        return super().get(request, *args, **kwargs)
-
-    def put(self, request, *args, **kwargs):
-        """
-        Atualizando um destino pelo `id`.
-        """
-        return super().put(request, *args, **kwargs)
-
-    def patch(self, request, *args, **kwargs):
-        """
-        Atualizando um destino pelo `id`.
-        """
-        return super().patch(request, *args, **kwargs)
-
-    def delete(self, request, *args, **kwargs):
-        """
-        Deletando um destino pelo `id`.
-        """
-        return super().delete(request, *args, **kwargs)
-
-
-class ListCreateDesination(ListCreateAPIView):
-    queryset = Destination.objects.all()
-    serializer_class = DestinationSerializer
-    parser_classes = [MultiPartParser]
-
-    def get(self, request, *args, **kwargs):
-        """
-        Listando os destinos.
-        """
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        """
-        Criando um destino.
-        """
-        return super().post(request, *args, **kwargs)
-
-
 rud_post = RetrieveUpdateDestroyPost.as_view()
 lc_post = ListCreatePost.as_view()
 post_home = PostHome.as_view()
-
-
-rud_destination = RetrieveUpdateDestroyDestination.as_view()
-lc_destination = ListCreateDesination.as_view()
