+++ pyproject.toml
@@ -55,9 +55,15 @@ exclude = '''
 line-length = 120
 
 select = [
-    "E",
-    "F",
-    "I",
+    #"E",
+    #"F",
+    #"I",
+    "E",  # pycodestyle errors
+    "W",  # pycodestyle warnings
+    "F",  # pyflakes
+    "I",  # isort
+    "C",  # flake8-comprehensions
+    "B",  # flake8-bugbear
 ]
 
 exclude = [
